refactor(verktoy): extract shared v2 calculator primitives to calcFields
Move the generic Info/Field/Seg/Stat building blocks and the no-NO
number formatters out of YieldCalculator into a shared calcFields module
so the next calculator can reuse them without duplication. YieldCalculator
now imports them; no behavioural change.
